requirement.txt
langchain
langchain_community
ipykernel
python-dotenv
pypdf
bs4
langchain-text-splitters
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_groq
langchain_core
fastapi
uvicorn
langserve
streamlit